+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Music/Soundcloud.cs
@@ -10,6 +10,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Music
 {
     public class Soundcloud : KaguyaBase
     {
+        [PremiumCommand]
         [MusicCommand]
         [Command("Soundcloud")]
         [Alias("sc")]