+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Premium/AddReaction.cs
@@ -10,6 +10,7 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Premium
 {
     public class AddReaction : KaguyaBase
     {
+        [FunCommand]
         [PremiumCommand]
         [Command("React")]
         [Summary("Takes a line of text and message ID and adds a reaction to the message. If no message ID is specified, the command-invoking " +