+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Kaguya Premium is now the only thing able to be purchased from the online store.
 - Fixed a bug with $nsfw bomb where it would break if tags were appended.
 - Fixed a bug where users would not receive their Kaguya Premium role automatically.
+- All Kaguya Premium commands are now displayed in their respective command categories, rather than on a separate command page list. All premium commands will have a `{$}` appended to the end of them, after the aliases.
 
 ### Version 2.3.2
 - Added lots of redundancy against crashes that occur in relation to being unable to send users direct messages. (This has gone on for too long!!)